refactor: type 수정

+++ backend/src/main/java/com/yagubogu/member/dto/v1/VictoryFairyProfileResponse.java
@@ -3,7 +3,7 @@ package com.yagubogu.member.dto.v1;
 import com.yagubogu.stat.dto.VictoryFairySummaryParam;
 
 public record VictoryFairyProfileResponse(
-        Integer ranking,
+        Long ranking,
         Integer rankWithinTeam,
         Double score
 ) {
